Move the test to CodeGenCXX
